requirements.txt
numpy
pandas
langchain>=0.1.0
pypdf
faiss-cpu
sentence-transformers
nltk
langchain-community
python-dotenv
langchain_openai
pymupdf
streamlit
langchain_huggingface
chromadb
